refactor(core): migrate ticket enums to own class
ref: #292
