+++ app/core/tests/unit/ticket_comment/test_ticket_comment_permission.py
@@ -16,7 +16,7 @@ from app.tests.abstract.model_permissions import ModelPermissions, ModelPermissi
 
 from project_management.models.projects import Project
 
-from core.models.ticket.ticket import Ticket, RelatedTickets
+from core.models.ticket import Ticket, RelatedTickets
 from core.models.ticket.ticket_comment import TicketComment
